tests/wpt/web-platform-tests/import-maps/acquiring/worker-request.html

<!DOCTYPE html>
<html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
const log = [];
new Worker('../resources/empty.js?pipe=trickle(d1)', {type: "module"});
</script>
<script type="importmap">
{
  "imports": {
    "../resources/log.js?pipe=sub&name=A": "../resources/log.js?pipe=sub&name=B"
  }
}
</script>
<script>
promise_test(() => {
  return import("../resources/log.js?pipe=sub&name=A")
    .then(() => assert_array_equals(log, ["log:B"]))
  },
  'After module worker creation import maps are still effective');
</script>
